[Destiny] First Commit of new cog
This is the first commit for the Destiny 2 Cog for Red Bot V3. This cog uses full OAuth 2 to access users information and stats in game. The Owner will need to provide API credentials then any user who wishes to use the commands will be asked to provide access to the bot under the provided scope then they can see their character information. There is a basic search for items which does not require OAuth and I plant to add more search options in the future.
There is a command to see Xur's wares when he's around and eververse items currently available.