[ReTrigger] Version 2.0.0
Remove option for `allowmultiple` to. This has been replaced with a new multi response trigger and payload to define one regex pattern for multiple responses allowing you to delete messages and add/remove roles, kick or ban.

This also adds support for customcommand like arguments inside text responses so you can do `[p]retrigger text test \btest\b Welcome to {guild.name}!` and it will replace the text with the servers current name. The same options work for commands as well so you can define the user or channel as an argument.

This also adds a new `mock` option which allows administrators to define a command to be run on their behalf. This can be very dangerous so it is owner only at the moment and should only be used to perform automated actions in the most extreme cases.