[Destiny] 1.1.0
Add History command to display last 5 activity types for each character. This comes with a gambit, raids, quickplay, and pvp command to quickly see these activities.

Optimize xur's wares to be one page and provide exactly when Xur is around next if he's not available. Show item perks as well.

Allow some commands to search other players stats such as loadout.

Fix a bug with loadout commands when a user has a title.