[EventPoster] 1.3.0
Add the ability to specify your chosen class when joining an event. Useful to let people know what class you're planning to play as when you join an event and even set a default class to always use.

Allow people to join events by reacting to emojis and easily leave and end events.